THUND-174: Default to immediate initialization of the association
