+++ lib/large_text_field/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module LargeTextField
-  VERSION = "1.3.0"
+  VERSION = "1.4.0"
 end
